17/9

+++ demo/src/main/java/com/example/demo/domain/CartDetail.java
@@ -15,7 +15,7 @@ public class CartDetail {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private int quantity;
+    private long quantity;
 
     private double price;
 
@@ -35,11 +35,11 @@ public class CartDetail {
         this.id = id;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(long quantity) {
         this.quantity = quantity;
     }
